project/requirements.txt
nextcord>=2.6.0  # Основная библиотека для работы с Discord API
python-dotenv==1.0.0
sqlalchemy==2.0.25
fpdf2==2.7.7
python-dateutil==2.9.0
aiohttp==3.9.3
reportlab>=3.6.0 # библиотека для эмоджи и вложений
requests
Pillow
